// services/tmc/src/shared-module/iframe-protocol-types.ts
import { ExerciseTaskGrading, ExerciseTaskSubmission, RepositoryExercise } from "./bindings"

/**
 * from: IFrame
 *
 * to: parent
 */
export interface CurrentStateMessage {
  message: "current-state"
  data: unknown
  valid: boolean
}

/**
 * from: IFrame
 *
 * to: parent
 */
export interface HeightChangedMessage {
  message: "height-changed"
  data: number
}

/**
 * from: IFrame
 *
 * to: parent
 */
export interface ReadyMessage {
  message: "ready"
}

/**
 * from: parent
 *
 * to: IFrame
 */
export type SetStateMessage = { message: "set-state" } & IframeState

export type IframeState =
  | {
      view_type: "exercise"
      exercise_task_id: string
      data: {
        public_spec: unknown
        previous_submission: ExerciseTaskSubmission | null
      }
    }
  | {
      view_type: "view-submission"
      exercise_task_id: string
      data: {
        grading: ExerciseTaskGrading | null
        user_answer: unknown
        public_spec: unknown
        model_solution_spec: unknown
      }
    }
  | {
      view_type: "exercise-editor"
      exercise_task_id: string
      repository_exercise: RepositoryExercise | null
      data: { private_spec: unknown }
    }